// src/frontend/components/Settings/types.ts
export interface BaseWidgetSettings<T = Record<string, unknown>> {
  enabled: boolean;
  config: T;
}
// ...
export interface PitlaneHelperWidgetSettings extends BaseWidgetSettings {
  config: {
    showMode: 'approaching' | 'onPitRoad';
    approachDistance: number;
    enablePitLimiterWarning: boolean;
    enableEarlyPitboxWarning: boolean;
    earlyPitboxThreshold: number;
    showPitlaneTraffic: boolean;
    showSpeedBar: boolean;
    showPositionBar: boolean;
    positionBarOuterThreshold: number;
    positionBarInnerThreshold: number;
    positionBarTargetZone: number;
    background: { opacity: number };
    progressBarOrientation: 'horizontal' | 'vertical';
    showPitExitInputs: boolean;
    pitExitInputs: {
      throttle: boolean;
      clutch: boolean;
    };
    showInputsPhase: 'atPitbox' | 'afterPitbox' | 'always';
  };
}